// src/lib/api/generated/models/bid.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */

export interface Bid {
  amount: number;
  auction_item_id: number;
  /** @nullable */
  auto_bid_max?: number | null;
  bid_time: string;
  bidder_id: number;
  created_at: string;
  id: number;
  is_winning: boolean;
  updated_at: string;
}